move test as its linux specific
